キャスト変換が必須でしたクソ

+++ ex05/ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: hshimizu <hshimizu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/15 13:46:23 by hshimizu          #+#    #+#             */
-/*   Updated: 2023/03/16 15:51:32 by hshimizu         ###   ########.fr       */
+/*   Updated: 2023/03/16 20:38:32 by hshimizu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,9 +27,9 @@ unsigned int	ft_strlcat(char *dest, char *src, unsigned int size)
 
 	len_dest = ft_strlen(dest);
 	len_src = ft_strlen(src);
-	if (size <= len_dest)
+	if ((int)size <= len_dest)
 		return (size + len_src);
-	while (*src && size-- > len_dest + 1)
+	while (*src && (int)size-- > len_dest + 1)
 		*(dest++ + len_dest) = *src++;
 	*(dest + len_dest) = '\0';
 	return (len_dest + len_src);
